Reduce memory overhead by changing size_t to int

+++ include/database.h
@@ -23,8 +23,8 @@ class Database {
 
   // Construct graph by frequent labels
   void construct_graphs(
-    const unordered_map<size_t, std::vector<size_t> > &frequent_vertex_labels,
-    const unordered_map<size_t, size_t> &frequent_edge_labels,
+    const unordered_map<int, std::vector<int> > &frequent_vertex_labels,
+    const unordered_map<int, int> &frequent_edge_labels,
     vector<Graph> &graphs);
 
   ~Database() {
@@ -37,7 +37,7 @@ class Database {
  private:
   static Database *instance_;
   vector<vector<string> > input_;
-  size_t num_graph_;
+  int num_graph_;
 };
 
 }  // namespace gbolt
